+++ tests/bincode.rs
@@ -9,7 +9,7 @@ proptest! {
     fn secretkey_serialization(
         bytes in prop::array::uniform32(any::<u8>()),
     ) {
-        let sk_result_from = SigningKey::<SpendAuth>::try_from(bytes);
+        let sk_result_from = SigningKey::<SpendAuth>::from_bytes(&bytes);
         let sk_result_bincode: Result<SigningKey::<SpendAuth>, _>
             = bincode::deserialize(&bytes[..]);
 
@@ -26,12 +26,12 @@ proptest! {
                 assert_eq!(&bytes[..], &bytes_bincode[..]);
 
                 // Check 3: From encoding should match original bytes.
-                let bytes_from: [u8; 32] = sk_bincode.into();
+                let bytes_from: [u8; 32] = sk_bincode.to_bytes();
                 assert_eq!(&bytes[..], &bytes_from[..]);
             }
             // Both agree on failure
             (Err(_), Err(_)) => {},
-            _ => panic!("bincode and try_from do not agree"),
+            _ => panic!("bincode and from_bytes do not agree"),
         }
     }